FIX: post comments 'x-access-token' exception

+++ routes/posts.js
@@ -264,6 +264,9 @@ router.delete('/:post_id/comments/:comment_id', async (req, res, next) => {
     try {
         const { email, password } = jwt.decode(token);
         const [[user]] = await pool.query(`SELECT id FROM users WHERE email = ? AND password = ?`, [email, password]);
+        if (!user) {
+            throw { status: 404, message: 'User not found' }
+        }
         const [rows] = await pool.query(`DELETE FROM post_comments WHERE user_id = ? AND post_id = ? AND id = ?`,[user.id, post_id, comment_id]);
         if (rows.affectedRows === 0)
             throw { status: 404, message: 'Comment not found' }
